+++ src/config/__init__.py
@@ -1,0 +1,3 @@
+﻿from src.config.config import Config
+
+__all__ = ['Config']
